requirements.txt
fastapi==0.95.0
uvicorn==0.21.1
aiohttp==3.8.1
pymysql==1.0.2
aiomysql==0.1.1
aiosmtplib==2.0.1
pysocks==1.7.1
websockets==10.4
bcrypt==3.2.2
discord_oauth2==0.2
steamsignin==1.0.1
pillow==9.3.0
python-dateutil==2.8.2
python-multipart==0.0.5